Added IDX file reading.
We can now read the label and image files by disassembling the header
and reading each byte as the program commands us to.

Important for reading multi-byte values: Some values are stored in 4
bytes, so you have to read them as whole by doing bit operations.
Namely, for a 4 bit, you need to read the first byte, shift it to the
left by 24 bits, then read the next byte, shift it by 16, repead for 8
and repeat for 0, like this:
(File.read() << 24) | (File.read() << 16) | (File.read() << 8) |
(File.read())

Otherwise you'll just read a single byte and get gibberish. For values
that are held on only ONE byte, you needn't worry.