Added file reading
Our knowledgebase now can correctly parse it's own files.

To do that we're using the Scanner class, but unlike what we'd expect
we're not using NextDouble function. Instead, we simply use Next to
prune out whitespaces (like, say, " "), and then the entire string is
parsed to a double.

All in all, we're set to start fiddling with actually categorizing
features. The database has also been made in such a way that we can very
easily add and remove the number of features should our method prove...
ineffective.

Let's pray it doesn't explode.