Implemented DataExtractor class, restructured the program flow
Oh boy, this is a big update so let's tackle it bit by bit.
Specifically, let's start with how the program flows now and what was
changed to accomodate it.

First, Knowledge class is the crux of the program. You only need to
declare it by itself, and it'll deal with everything mostly on its own.
When declaring the Knowledgebase, you also declare a DataExtractor.

DataExtractor class is a new class that serves to run the feature
extraction algorithms from the image as well as read the labels, though
this is just a side-feature and it's hardly novelty. To do this
DataExtractor copies each image from the file read and runs a specified
(by us) feature finding algorithm.

Finally, to read the images, DataExtractor employs the ReadDatabase
functions from itself, which have been changed slightly to read a
specified NUMBER of data from the file instead of gorging the entirety
of it. It'll simply save us space so we don't waste time loading the
entire 60,000 boxes of madness.

So, how it goes.
Declare Knowledgebase with a SIZE.
Knowledgebase declares Extractor with size, specifies its own tables
with size.
Extractor specifies its own tables for holding the read data with size,
then calls Read to read the (size) amount of data from the files.

To-Do List from this point:
> Figure out the gap algorithm for feature 2, 4 and 6.
> Figure out where to put the IDX files so that the program sees them
without resorting to absolute paths.
> Figure out the k-NN algorithm for matching in DataMatcher class.

...which honestly isn't much. This is a good thing, since it means we're
-almost- there, huzzah!