Began work on restructurizing the program.
Will remove the direct necessity to nest classes with one another, so
that Knowledgebase doesn't necessarily come packaged with everything
else, as to avoid errors with erroneous construction.

Also changed the way we extract features by writing a dedicated function
for it so we don't have redundant code.

We'll finish working on it this weekend.