Restructurized the system and implemented full reading-extraction-writing functionality.
By that I mean, our program now works and calculates the densities
correctly. We're no longer chaining it all together, instead using
Extractors and Knowledgebases independently. As the Knowledgebase is
just a bank, our main "tinkering" function is the Extractor.

!! IT IS ABSOLUTELY IMPERATIVE THAT OUR SIZES ARE SET IN STONE BEFORE WE
ATTEMPT ANY CALCULATION OR TRICKERY. IF WE DON'T WE'RE GOING TO HAVE A
BAD TIME !!

The Extractor class had one major change, namely how we calculate the
feature from the image data. We calculate the amount of whitespace, and
only that, expanded to feature 16 (4x4) zones. This should give us about
92-94% success rate. Now, the only thing left on the to-do list is to
write a Matcher class that will use the Extractor to read the values and
match them against the labels. Easy, huh? (Famous last words)