Almost finished the Comparator Function
Almost implemented the kNN algorithm using the Manhattan Metric (or how
we call it, "metoda uliczna", since it's based off the streets of
Manhattan).

In essence, I added a way for our KnowledgeGathere to distinguish
between what sort of database are we filling - the knowledge or the
matcher.  When we have both databases filled (again, please make sure
they're both the correct size!), we can invoke the testNumbers(int k)
function, which in itself is the kNN algorithm, which works as follows:
> Calculate the distance between the object we're on and EVERY OTHER
OBJECT IN THE KNOWLEDGEBASE. You do this by summing the absolute values
of the differences between each individual features (aka sum |x_i -
y_i|).
> Sort these from the lowest to highest.
> Pick k lowest, i.e. the ones closest to our object.
> Sum up how many times have we encountered each label in that cluster.
> Decide which label is the most common.
> Assign that label.

And additionally, I implemented a very simple algorithm at the tail of
it all.
> If the label is the same as the original label, increase the success
rate by 1.
> Divide the success rate by the matchbase's size.
> Multiply by 100, add "%" at the end.
> Enjoy your percentage of correctness (sounds like a D&D item I swear).

"Almost" finished because the arrays gave me a headache. I'll take a
break and see if I can finish it today, but honestly I'm crapping arrays
out the side of my skull at the moment.