Finished Implementing the base matching algorithm, AGAIN.
Currently, with a k around 8 and half of the known test-set we're
achieving the correct guess rate of 82%. I'll write the entire 60k
database into a file and load it in by hand so we don't have to mull it
each and every time, and fiddle with the way we cound the int numbers
and assign labels in the rare situations where mutliple labels fit. If
we still don't break 90, well, I'll just slap on more density zones, and
it has to work :D

Also git has emotes.
:whale:
I have lost my toucan.