Finished the quest for 90%!
In actuality, we've scored 96% to be precise. :+1:

The solution was childishly simple. We've amped up our number of sectors
from 16 (4x4) to... well, 49 (7x7). Otherwise, it works pretty much as
before, as advertised, no changes to our model aside from amping up the
sector focus.

One could say we're dangerously close to per-pixel approximation,
simplified to small squares instead. It's a very primitive algorithm, a
very simple one and tends to take a bit of time, but it's also
stunningly effective since, you could say, it's essentially comparing
every picture, as is, pixel-wise, to the one we have and finds the label
most suiting.

Again, simple but memory-heavy. Huzzah! :smiley: