Finialized a proper Training Image reading function.
The function will correctly read our image based solely on its number in
the string we create from first inputting how many of each type of
texture we want to read.

In other words, we first determine texture amounts. Say, 20, 30, 50 and
100.
Then, we give the image number from this list we want to read (yes, this
means if we give 50 of stalt, it'll read the first 50 only, this is
intended).

The program will then determine what type of texture we're reading and
will properly  navigate to the correct file.