Finished writing the flat FFT Transformation.
Rewrote it from ground up. Curiously there's an IndexOutOfBounds error
for FFT Extraction when we approach the last number but... it yells at
us that we crossed 400 while the limit is 400 and the last image (that
we actually load) is #400... so we did pass the entire thing.

It's probably a small thing.

He'll probably bitch about *a* small thing later.
Mark my words, future self.