Added FFT Swapping for Filtering Process + Fixed Image Access
As with the old lore, you need to swap the FFT halves so that you get a
central star. On the bright side we won't have to bother much with
reconstructing the FFT image, since this is as far as we go in FFT
transformations.

Also fixed the FFT feature counter in the Knowldegebase, so that it's no
longer throwing me stupid errors. Thank some higher power that I fixed
it now and not when we started implementing feature list.