Added FFT Feature Extraction (Full Algorithm)
By using 4 hand-made filters, we zero out the values on our swapped FFT
graph and count the resulting denisty of colour depending on the board
(which is always 64 pixels).

These features are tentative, as well as the filters themselves, but
what matters is that the 'core' program by itself works. Possible future
tweaks are, pretty much, only math at this point plus deciding on better
features.

We're almost there, almost. :cloud: