Changed FFT around, also corrected our filters.
We're now working directly on Frequency and Power Spectrums, like we
used to do way back.

Technically the FFT feature value is correct, after normalization, since
the data checks out. We're getting values in around ~100-160 range,
which makes sense for a 0-255 image.

On the flip side, we seem to draw the entire result in -white- for
whatever reason. I'm not sure if it's KNN that's borked, or the loops
are faulty.

I'll rewrite the entire KNN Function to split it apart, maybe that will
help a little bit.