Finished implementing the Bayes method.
It doesn't work, but I don't care.

You see, Naive Bayes works by probability assumption. As per example I
linked earlier, it assumes the probability of, perfectly, entire ranges
of numbers. For example, it'll work very well if the feature we're
looking over has a range of binary, or up to 4. Ages, salary described
with words, agreement of yes or no to some statement, these things will
work perfectly with Naive Bayes.

About 1000+ double-type numbers? Nu-uh.
I mean, it is possible to consolidate our features in both cases so that
Bayes can read them by, say, rolling the values to 0-10 integers, but
you can imagine this essentially butchers the probability ratios.

In essence, KNN, despite it's tradiness proves to be the universal
method while Bayes, for this sort of task, is more cumbersome than
useful in the end.

Q.E.D., the program is finished. I am so done with this.