Modified the algorithm assignment method.
Strap on kids it's gonna get weird. I've changed how we fly through our
image based on the Profss's guide. Instead of skipping through 64x64
large squares, we now go bit by bit.

In other words, while we still use the default 64x64 square to match
(for the sake of matching FFT densities), only the top-left pixel is
considered aligned. We then move the square a bit to the right and
repeat the process for the next pixel. This does give us rather...
monstrous effects on the bottom-right edge but this is within the norm
of the algorithm (read: I can't be arsed to come up with a dynamic
allocation solution for our square size).

Next step: finding a way to get the Bayes algorithm to always catch
something, perhaps by calculating a range error and just slapping
pinging +1 whenever we get inside of it. Who knows?