Bayes Works.
So, I added the following changes to optimize it a bit:
A) Bayes doesn't compare flat numbers, instead it calculates the raw
distance between the numbers, then checks if the distance is within a
threshold to consider it a  'hit'.

B) We start our feature numbers from 1, not 0. Since we all added +1 to
everything in that sense it doesn't affect the probabilities at all
aside from removing null ones.

C) THERE WAS A TINY FUCKING ERROR IN ONE FUCKING LOOP THAT TOOK ME THREE
WEEKS TO FUCKING FIX FOR FUCK'S SAKE ALL OF THIS JUST OVER A SINGLE
BINARY DIGIT. IT WAS LITERALLY A CASE OF STARTING FROM 1, NOT FROM 0,
GOD DAMMIT.