Ironed out the kinks!
But the method requires a bit more fixing. Namely, it detects a single
grape two to three times, sometimes detecting emtpy space.

I need to increase the mid-grape distance as well as figure out a better
cross-checking algorithm, as in for checking if the box is nearby in a
cross(top-right, bottom-left, for example) fashion.