Added the colour calculation function and improved the general flow of the program.
We can now set the average size and the cutoff value (keep it low, at
like 2%) from the main.java level. The grapes count themselves but I
need to find a better way of doing that so that our results are less
ubiquotous.

...since it often mistakes purple for green.
Don't ask me how.

What matters is that it detects the grapes very well with a low-low
error if we set the values just right. I mean, it can sometimes mistake
a particularly large cluster of 3 grapes for 4 or 5 grapes, but that's
still pretty darn epic!