Added the shape-recognition algorithm (faulty!)
Okay so, this is how I solved this conundum. It obviously fails
spectacularly on tight clusters but hey, it's something, right?

In essence, whenever we detect a nut and we detect it's not a dark nut,
we go into shape detection.

Shape detection picks a pixel in the middle of the segment, on its right
edge. This is approximated to be the 'middle' of the nut considering
we're going from top-to-bottom, left-to-right, in that order. From that
point, we draw a circle the size of an average nut we decided to use.
The circle then counts the number of white pixels. If the ratio is lower
than 40.2% (by way of repeated testing), we declare the nut to be a
circular nut. If there's more whitness than that percentage cutoff, it
means we're actually touching an oblong nut.

Again, this will fail spectaccularly on tight clusters but for lone nuts
it works flawelssly. Happy hunting.